refactor: extract design system into MessengerDesign.swift
Move the shared design-system components (AvatarView/AvatarPalette,
Messages* backgrounds, GlassSeparator and the Glass* chrome, plus the
View styling modifiers) out of the 5.7k-line MessengerShellView.swift
into a dedicated MessengerDesign.swift. Pure code movement: top-level
declarations promoted from file-private to internal so cross-file
references resolve; no logic, signature, or behavior changes.

Build green; test suite unchanged (the two pre-existing image-downsampling
failures are unrelated to view extraction).
